requirements.txt
python-telegram-bot==20.1
openai>=0.26.1
PyYAML==6.0
pymongo==4.3.3
python-dotenv==0.21.1
revchatgpt
tiktoken
tls-client
requests
aiohttp